edgar-monitor/requirements.txt
feedparser
requests
python-dotenv
edgartools
pytest
networkx
